Isolated CostList into CostListService.
This is a pure refactoring commit (zero behavior changes).

CostList.Calculator was difficult to test because it was entangled with the logic
of ILabel and Recipe. This commit moves most of the logic from CostList
into AbstractCostListService, which is generic over arbitrary Labels,
Recipes, and CostLists. By adding this extra layer of indirection (which
should mostly be optimized away by the JVM), we can write tests that are
unconcerned about things like locales and ore dictionaries, and which
don't manipulate the global recipe list and so can't collide with one
another.

While I have rearranged things, I have avoided changing any of the fundamental
logic that previously existed in CostList. The new calculation code follows the
old calculation code line by line, differing only in its use of the indirection
layer.

MainCostListService is the shim that connects AbstractCostListService to
the concrete classes ILabel, Recipe, and CostList, so most code external
to the `structure` package should now go through `MainCostListService.INSTANCE`.
There are examples of this in `GuiCraft` and `JecaOverlayHandler`.